requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.8.1
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.2.0
autobahn==23.1.2
backcall==0.2.0
beautifulsoup4==4.11.2
bleach==6.0.0
cart==1.2.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
comm==0.1.2
cryptography==39.0.1
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
diff-match-patch==20200713
Django==5.0.4
django-admin-soft-dashboard==1.0.12
django-ckeditor==6.7.1
django-cors-headers==4.3.1
django-import-export==3.2.0
django-jazzmin==3.0.0
django-js-asset==2.2.0
django-object-tools==2.0.0
django-order==0.0.7
django-phonenumber-field==7.1.0
django-shopping-cart==0.1
djangorestframework==3.15.1
dnspython==2.3.0
email-validator==2.0.0.post2
et-xmlfile==1.1.0
executing==1.2.0
fastapi==0.97.0
fastjsonschema==2.16.2
fqdn==1.5.1
frozenlist==1.3.3
h11==0.14.0
httpcore==0.17.2
httptools==0.5.0
httpx==0.24.1
hyperlink==21.0.0
idna==3.4
ipykernel==6.21.2
ipython==8.10.0
ipython-genutils==0.2.0
ipywidgets==8.0.4
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.6.1
jupyter-events==0.6.3
jupyter-server-proxy==3.2.2
jupyter_client==8.0.3
jupyter_core==5.2.0
jupyter_server==2.3.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.5
MarkupPy==1.14
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
mistune==2.0.5
multidict==6.0.4
mysqlclient==2.2.4
nbclassic==0.5.2
nbclient==0.7.2
nbconvert==7.2.9
nbformat==5.7.3
nest-asyncio==1.5.6
notebook==6.5.2
notebook_shim==0.2.2
numpy==1.24.2
odfpy==1.4.1
openpyxl==3.1.2
orjson==3.9.1
packaging==23.0
pandocfilters==1.5.0
parso==0.8.3
phonenumberslite==8.13.11
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==3.0.0
prometheus-client==0.16.0
prompt-toolkit==3.0.37
psutil==5.9.4
psycopg2==2.9.9
pure-eval==0.2.2
pycparser==2.21
pycryptodome==3.17
pydantic==1.10.9
Pygments==2.14.0
PyMySQL==1.1.0
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.7
python-multipart==0.0.6
pywin32==305
pywinpty==2.0.10
PyYAML==6.0
pyzmq==25.0.0
qtconsole==5.4.0
QtPy==2.3.0
razorpay==1.3.0
requests==2.28.2
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
Send2Trash==1.8.0
simpervisor==0.4
six==1.16.0
sniffio==1.3.0
soupsieve==2.4
sqlparse==0.4.3
stack-data==0.6.2
starlette==0.27.0
tablib==3.4.0
terminado==0.17.1
tinycss2==1.2.1
tornado==6.2
traitlets==5.9.0
txaio==23.1.1
typing_extensions==4.6.3
tzdata==2022.7
ujson==5.8.0
uri-template==1.2.0
urllib3==1.26.15
uvicorn==0.22.0
vpython==7.6.4
watchfiles==0.19.0
wcwidth==0.2.6
webcolors==1.12
webencodings==0.5.1
websocket-client==1.5.1
websockets==11.0.3
widgetsnbextension==4.0.5
xlrd==2.0.1
xlwt==1.3.0
yarl==1.8.2
